feat: 두 클래스 하나로 통합
